pyproject.toml
[tool.poetry]
name = "apps"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.8"
django-ninja = "^1.1.0"
django-environ = "^0.11.2"
psycopg = "^3.1.18"


[tool.poetry.group.dev.dependencies]
black = "^24.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
